% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centroid_scores.R
\name{calculate_centroid_scores}
\alias{calculate_centroid_scores}
\title{Calculate Centroid-Based Modules}
\usage{
calculate_centroid_scores(x, input_species = "human")
}
\arguments{
\item{x}{a matrix, of Ensembl IDs (p) x samples (n), already log-transformed}

\item{input_species}{character string, "human" or "mouse" (default = "human")}
}
\value{
a matrix of module scores (p) x samples (n)
}
\description{
This function takes in a data matrix of Entrez IDs x samples and returns module scores.
}
