% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_median_modules.R
\name{calculate_median_modules}
\alias{calculate_median_modules}
\title{Calculate Median-Based Modules}
\usage{
calculate_median_modules(
  x,
  filter_genes = TRUE,
  filter_gene_thresh = 0.7,
  row_median_center = TRUE,
  column_standardize = TRUE,
  custom_median_modules = NULL,
  input_species = "human"
)
}
\arguments{
\item{x}{a matrix, of Ensembl IDs (p) x samples (n), already log-transformed (if applicable)}

\item{filter_genes}{logical, whether to filter genes prior to median module calculation (default = TRUE)}

\item{filter_gene_thresh}{a number between 0 and 1, fraction of samples each gene must be detected in to pass filter. This parameter does nothing if filter_genes is set to FALSE (default = 0.7)}

\item{row_median_center}{logical, whether to median-center rows prior to median module calculation (default = TRUE)}

\item{column_standardize}{logical, whether to column-standardize data matrix prior to median module calculation (default = TRUE)}

\item{input_species}{character string, "human" or "mouse" (default = "human")}
}
\value{
a matrix of module scores (p) x samples (n)
}
\description{
This function takes in a data matrix of Ensembl IDs x samples and returns module scores.
}
