% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_modules.R
\name{calculate_modules}
\alias{calculate_modules}
\title{Calculate Breast Cancer Modules curated in the Perou Lab}
\usage{
calculate_modules(
  x,
  as.is = FALSE,
  standardize_output = TRUE,
  filter_genes = TRUE,
  filter_gene_thresh = 0.7,
  row_median_center = TRUE,
  column_standardize = TRUE,
  log_transform = TRUE,
  include_median_modules = TRUE,
  include_single_genes = TRUE,
  include_special_models = TRUE,
  include_prognostic = TRUE,
  custom_median_modules = NULL,
  input_species = "human"
)
}
\arguments{
\item{x}{a matrix, preferably of Ensembl IDs (p) x samples (n)}

\item{standardize_output}{logical, whether to row standardize the output (default = TRUE)}

\item{filter_genes}{logical, whether to filter genes prior to median module calculation (default = TRUE)}

\item{filter_gene_thresh}{a number between 0 and 1, fraction of samples each gene must be detected in to pass filter. This parameter does nothing if filter_genes is set to FALSE (default = 0.7)}

\item{row_median_center}{logical, whether to median-center rows prior to median module calculation (default = TRUE)}

\item{column_standardize}{logical, whether to column-standardize data matrix prior to median module calculation (default = TRUE)}

\item{log_transform}{logical, whether to log-transform data matrix prior to median module calculation (default = TRUE)}

\item{include_median_modules}{logical, whether to calculate median-based modules (default = TRUE)}

\item{include_single_genes}{logical, whether to calculate single gene modules (default = TRUE)}

\item{include_special_models}{logical, whether to calculate centroid-based modules and special models (default = TRUE)}

\item{include_prognostic}{logical, whether to include prognostic models (default = TRUE)}

\item{custom_median_modules}{list of vectors (optional), e.g. list("CustomModule1" = c("GeneA", "GeneB", "GeneC"), "CustomModule2" = c("GeneD", "GeneE", "GeneF"))}

\item{input_species}{character string, "human" or "mouse" (default = "human")}
}
\value{
a matrix of module scores (p) x samples (n)
}
\description{
This function takes in a data matrix of genes x samples and returns module scores.
}
