% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_single_genes.R
\name{calculate_single_genes}
\alias{calculate_single_genes}
\title{Calculate Single Gene Modules}
\usage{
calculate_single_genes(x, input_species = "human")
}
\arguments{
\item{x}{a matrix, of Ensembl IDs (p) x samples (n), already log-transformed}

\item{input_species}{character string, "human" or "mouse" (default = "human")}

\item{single_genes}{a vector of Gene Symbols}
}
\value{
a matrix of module scores (p) x samples (n)
}
\description{
This function takes in a data matrix of Ensembl IDs x samples and returns module scores.
}
