% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_special_models.R
\name{calculate_special_models}
\alias{calculate_special_models}
\title{Calculate Special Model Modules}
\usage{
calculate_special_models(x, input_species = "human")
}
\arguments{
\item{x}{a matrix, of Entrez IDs (p) x samples (n), already log-transformed}

\item{input_species}{character string, "human" or "mouse" (default = "human")}

\item{special_model_functions}{vector of files, with each file containing the source function for one special model}
}
\value{
a matrix of module scores (p) x samples (n)
}
\description{
This function takes in a data matrix of Entrez IDs x samples and returns module scores.
}
