% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_duplicates.R
\name{collapse_duplicates}
\alias{collapse_duplicates}
\title{Collapse duplicates row}
\usage{
collapse_duplicates(x, ids = row.names(x), sum_or_mean = "sum")
}
\arguments{
\item{x, }{a matrix of features x samples}

\item{ids, }{a vector of ids (same length as features) to collapse by}

\item{sum_or_mean, }{a string, method to collapse rows (default = "sum")}
}
\value{
a matrix of collapsed features x samples
}
\description{
This function takes in matrix and collapses the rows by the \code{ids}. This is used for when e.g. multiple probes map to the same gene, and those values (rows) need to be aggregated.
}
