% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_ensembl.R
\name{convert_to_ensembl}
\alias{convert_to_ensembl}
\title{Convert Data Matrix to Ensembl Matrix}
\usage{
convert_to_ensembl(x, species = "human", sum_or_mean = "sum")
}
\arguments{
\item{x, }{a matrix of numbers, rows = genes and columns = samples}

\item{species, }{character string, "human" or "mouse" (default = "human")}

\item{sum_or_mean, }{a string, method to collapse rows (default = "sum")}
}
\value{
a matrix of mapped Ensembl IDs x samples
}
\description{
This function takes in a matrix of genes x samples and returns a matrix of mapped Ensembl IDs x samples.
}
