% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_to_ensembl.R
\name{entrez_to_ensembl}
\alias{entrez_to_ensembl}
\title{Convert Entrez IDs to Ensembl IDs}
\usage{
entrez_to_ensembl(entrez_ids, species = "human")
}
\arguments{
\item{entrez_ids}{a vector of numbers or numbers as characters}

\item{species}{character string, "human" or "mouse" (default = "human")}
}
\value{
a vector of mapped Ensembl IDs
}
\description{
This function takes in a vector of Entrez IDs and return a vector of mapped Ensembl IDs.
}
