% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modernize_entrez_ids.R
\name{modernize_entrez_ids}
\alias{modernize_entrez_ids}
\title{Modernizes Entrez IDs}
\usage{
modernize_entrez_ids(entrez_ids, species = "human")
}
\arguments{
\item{entrez_ids}{a vector of Entrez IDs}

\item{species}{character string, "human" or "mouse" (default = "human")}
}
\value{
a vector of "modernized" Entrez IDs
}
\description{
This function takes in vector of Entrez IDs and returns a vector of updated Entrez IDs based on NCBI's \code{gene_history.gz} file.
}
