% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_for_median_modules.R
\name{normalize_for_median_modules}
\alias{normalize_for_median_modules}
\title{Performs normalization procedure for median-based modules}
\usage{
normalize_for_median_modules(
  x,
  filter_genes = TRUE,
  filter_gene_thresh = 0.7,
  row_median_center = TRUE,
  column_standardize = TRUE
)
}
\arguments{
\item{x}{a matrix, already log-transformed or similar}

\item{filter_genes}{logical, whether to filter genes prior to median module calculation (default = TRUE)}

\item{filter_gene_thresh}{a number between 0 and 1, fraction of samples each gene must be detected in to pass filter. This parameter does nothing if filter_genes is set to FALSE (default = 0.7)}

\item{row_median_center}{logical, whether to median-center rows prior to median module calculation (default = TRUE)}

\item{column_standardize}{logical, whether to column-standardize data matrix prior to median module calculation (default = TRUE)}
}
\value{
a matrix of normalized values
}
\description{
This function normalizes a data matrix
}
