% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_modules.R
\name{refresh_modules}
\alias{refresh_modules}
\title{Refresh module database}
\usage{
refresh_modules()
}
\description{
This function regenerates the \code{median_modules.RDS} and \code{all_centroids.RDS} files to be loaded in their respective functions
}
