% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_to_ensembl.R
\name{symbol_to_ensembl}
\alias{symbol_to_ensembl}
\title{Convert Gene Symbols to Ensembl IDs}
\usage{
symbol_to_ensembl(gene_symbols, species = "human")
}
\arguments{
\item{gene_symbols}{a vector of characters}

\item{species}{character string, "human" or "mouse" (default = "human")}
}
\value{
a vector of mapped Ensembl IDs
}
\description{
This function takes in a vector of Gene Symbols and return a vector of mapped Ensembl IDs.
}
